chore: shellcheck

+++ debug.sh
@@ -1,24 +1,14 @@
-#!/usr/bin/env bash
+#!/bin/bash
+set -ueo pipefail
 
 # https://raw.githubusercontent.com/JamesKyburz/docker-node-debug/master/debug.sh
 
-service=$1
+log_error() { echo -e "\033[0m\033[1;91m${*}\033[0m"; }
+log_success() { echo -e "\033[0m\033[1;92m${*}\033[0m"; }
 
-declare -r cs_color_green="\033[1;92m"
-declare -r cs_color_red="\033[0;91m"
-declare -r cs_color_reset="\033[0m"
+service="${1:-}"
 
-alias log_success="log green "
-alias log_error="log red "
-
-shopt -s expand_aliases
-
-log () {
-  local log_color="cs_color_${1}"
-  echo -e "${cs_color_reset}${!log_color}${*:2}${cs_color_reset}"
-}
-
-if [ -z $service ]; then
+if [[ -z "${service:-}" ]]; then
   log_error "parameter 1 must be service name"
   log_error "optional parameter 2 is node working directory"
   log_error "optional parameter 3 is entry point for node"
@@ -32,37 +22,35 @@ ymlfiles=${DOCKER_COMPOSE_YAML_FILES:--f docker-compose.yml -f docker-compose.ov
 
 compose="docker-compose $ymlfiles"
 
-function debugservice {
-  $compose -f docker-compose.debug.yml up -d --build $service
+function debugservice() {
+  $compose -f docker-compose.debug.yml up -d --build "${service:?}"
 }
 
-function removeservice {
-  $compose kill $service
-  $compose rm -f $service
+function removeservice() {
+  $compose kill "${service:?}"
+  $compose rm -f "${service:?}"
 }
 
 removeservice
 
-cat << EOF > docker-compose.debug.yml
+cat <<EOF >docker-compose.debug.yml
 version: '3'
 services:
-  $service:
-    entrypoint: sh -c "node --inspect-brk=0.0.0.0:9229 $workdir$nodeindex"
+  ${service:?}:
+    entrypoint: sh -c "node --inspect-brk=0.0.0.0:9229 ${workdir:?}${nodeindex:?}"
     command: []
-    environment:
-      - RABBITMQ_PREFETCH=1
     ports:
-      - "$port:9229"
+      - "${port:?}:9229"
 EOF
 
 debugservice
 
-log_success "open chrome with url chrome://inspect, then select inspect for $service"
+log_success "open chrome with url chrome://inspect, then select inspect for ${service:?}"
 
 while true; do
   log_success "press any key to stop debugging press (r to restart debug)"
-  read -n 1 key
-  if [ "$key" == "r" ]; then
+  read -n 1 -r key
+  if [ "${key:-}" == "r" ]; then
     removeservice
     debugservice
   else
@@ -73,4 +61,4 @@ rm -rf docker-compose.debug.yml
 echo ""
 log_success "restarting $service without debug"
 removeservice
-$compose up -d --build $service
+$compose up -d --build "${service:?}"
